k8s-base: Move etcd members cleanup to separate role
Executing the etcd clean up tasks inside the k8s-master role does not
properly clean up the etcd members if master-0 got removed. The reason
for that is, that the k8s-master role is executed on all control plane
nodes, but at first (refer serial flag here [1]) on the master-0. Here
is what happens: After old-master-0 got destroyed, stage3 is executed.
As the k8s-master role is executed at first on the new-master-0 and that
ones a fresh node which is not part of the Kubernetes cluster yet, it
has no kubeconfig and can not clean up the etcd members list and
new-master-0 tries to join the cluster which then fails as the etcd
members list did not get cleaned up prior to its join attempt.

Moving the tasks to a separate role which is executed in serial on all
control plane nodes ensures that the etcd members list gets cleaned up
as it is ensured that the tasks get executed on a control plane node
which is part of the k8s cluster before new-master-0 tries to join the
k8s cluster(and if there is no node left, then it is totally fine that
new-master-0 inits a new cluster).

[1] https://gitlab.com/yaook/k8s/-/blob/06e8daa30c6f9c0a412ba79848ba3116fbc5c7d8/k8s-base/install_k8s_base.yaml#L144