OS-Cleanup: Process outcome of deletion API calls
This commit extends the OpenStack project clean up script such that the
return code of a resource deletion API call is printed and the script
raises an Exception if the deletion of a resource failed. It is also
ensured that the OpenStack connection is properly closed on failures.

Unfortunately, not every call provides a return code.