OS-Cleanup: Explicitly pass IDs instead of objects
It is weird that the deletion of resources may not work when the script
is triggered inside a docker container. Please refer to !549 [1].

This commit adjusts the OpenStack project clean up script such that the
IDs of a resource are explicitly passed to the deletion functions and
therefore to the API calls. Hopefully, this will prevent further
ambivalent behavior.

[1] https://gitlab.com/yaook/k8s/-/merge_requests/549