Integrate Vault (as workload)
This allows to run a HashiCorp Vault instance inside a yaook/k8s cluster
easily.

This is a `git subtree` import of the `vault_v1` role from
https://gitlab.com/yaook/incubator/management-cluser, with slight
modifications to integrate well in yaook/k8s.

Note that this is not yet production ready. Missing bits are high
availability and backups.