kubeadm: Fix node join attempts with taints
This commit makes the kubeadm join configuration template consistent
across the k8s-master and k8s-worker role regarding the taints
templating. Initializing a k8s cluster with a tainted node has been
fixed in !510. If there are problems in the future, we may want to go
for the solution presented in !556.

Without this patch, you get greeted with a nice error when attempting to
join a node with predefined taints:

```
fatal: [managed-k8s-master-2]: FAILED! => changed=true
  cmd: umask 077 && kubeadm join --config=/tmp/kubeadm-join-config.yaml
  delta: '0:00:00.075057'
  end: '2022-06-16 09:15:58.467417'
  msg: non-zero return code
  rc: 1
  start: '2022-06-16 09:15:58.392360'
  stderr: |-
    could not interpret GroupVersionKind; unmarshal error: error converting YAML to JSON: yaml: line 14: did not find expected node content
    To see the stack trace of this error execute with --v=5 or higher
  stderr_lines: <omitted>
  stdout: ''
  stdout_lines: <omitted>
```