k8s-config: Add rook version mapping for k8s v1.21 and v1.22
Seems like it has been forgotten for v1.21. Removing these can be
discussed in #481. As there are clusters out there which still need an
update to rook v1.6.7, I am scared to bump the rook version here.

If it is not defined here, the user has to set it in his config.toml
(mandatory variable) which we do want to avoid.

https://gitlab.com/yaook/k8s/-/issues/481