Connect-OS: Split out kube-controller patching tasks
Prior to updating Kubernetes to version v1.22, it is essential to update
the OpenStack cloud provider integration components because they are
otherwise unable to properly talk to the Kubernetes API and the
OpenStack connection is broken which leads to no-fun functional
disruptions.

This commit moves the kube-controller patch to an own file. It is only
needed for component versions we used prior to k8s v1.22.