connect-k8s-to-openstack: Make components k8s version dependent
This commit updates the Kubernetes-on-OpenStack integration role
"connect-k8s-to-openstack" such that the setup procedure is adjusted to
apply manifests dependent on the currently deployed Kubernetes version.
Templates are therefore stored in accordance to their associated/mapped
Kubernetes version [1]. Although we use the same versions for k8s v1.17
to v1.21, we need this in preparation for the Kubernetes update to v1.22
[2] which requires updating these components to recent versions as some
APIs will be stopped being served.

[1] https://gitlab.com/yaook/k8s/-/blob/c1b2b6afc99ba02fcf925a300cc14eb8241daf73/k8s-base/roles/k8s-config/defaults/main.yaml
[2] https://gitlab.com/yaook/k8s/-/merge_requests/506/