connect-k8s-to-openstack: Update components for k8s v1.22
This commit updates all components which are needed for the integration
of Kubernetes on OpenStack as cloud provider. This includes the CCM
(Cloud Controller Manager) and the Cinder CSI (Container Storage
Interface). Links to the releases are given in the k8s-config role [0]
where these versions are mapped to the Kubernetes version. The templates
are based upon the ones provided in the official repository [1][2]. The
setup procedure already has been made compliant to the official docs of
the CCM [3] and Cinder CSI plugin [4].

We MUST update these components prior to the update to Kubernetes v1.22
as that Kubernetes version stops serving some APIs which are used by the
old versions which will then lead to a broken OpenStack integration.

[0] https://gitlab.com/yaook/k8s/-/blob/8973c33350420bb172d750cb549e4e3686a03607/k8s-base/roles/k8s-config/defaults/main.yaml
[1] https://github.com/kubernetes/cloud-provider-openstack/tree/8ea8cfb8321246d3f05fcbedab668aa330028575/manifests/controller-manager
[2] https://github.com/kubernetes/cloud-provider-openstack/tree/8ea8cfb8321246d3f05fcbedab668aa330028575/manifests/cinder-csi-plugin
[3] https://github.com/kubernetes/cloud-provider-openstack/blob/8ea8cfb8321246d3f05fcbedab668aa330028575/docs/openstack-cloud-controller-manager/using-openstack-cloud-controller-manager.md
[4] https://github.com/kubernetes/cloud-provider-openstack/blob/8ea8cfb8321246d3f05fcbedab668aa330028575/docs/cinder-csi-plugin/using-cinder-csi-plugin.md