k8s-v1.22-update: Include OS-integration components updates
Not sure if this has been mentioned before, but in advance to the update
to Kubernetes v1.22 we need to update the cloud provider (in our case
OpenStack) integration components as some APIs will be stopped being
served which are used in the old versions of these components.

Therefore, this commit adds an update tasks file to the connect-k8s-to-
openstack role which is included during the upgrade to k8s v1.22
procedure/playbook. The components (CCM / CSI plugin) are updated
before the actual Kubernetes update to have the minimum possible
disruption.