docker: reload/restart during config change
This commit makes ansible reload the docker daemon service unit after a
config change even if the runtime is already under load. Reload isn't
harmful because it doesn't restart containers. The previous "restart
upon a change if the runtime is not under load"-behavior remains
untouched.

The new reload/restart handler helps with making non-disruptive
amendments to the daemon.json.