update_inventory: refer to template in error case
When update_inventory.py detects that a mandatory variable was not
declared in the config.toml, its error message does not provide the full
context. If, e.g., `A.B.C` is missing, then the user is only warned
about the lack of `.C` but is not informed where `.C` should actually be
embedded. I am too lazy right now to fix this in tooling so I only
advise the user to check the template manually for more information.