etcd-backup: Change location of default S3 config
The purpose of `<role>/templates` is to store templates which are
instantiated by that very role. It is not meant as a placeholder for
templates which should be copied manually by the user. Our `templates/`
folder in the root directory is a better spot.

Regarding the naming: `config.yaml` is not a good default because we
already have a `config.toml` in config/ which can easily lead to
confusion. The file name should reflect the purpose of its contents.