vault: enable Raft (Integrated Storage) backend for dev server
This allows to snapshot the data and import it to a later productive
vault cluster when bootstrapping a "self-hosted" cluster.

Note that this does *not* allow to merge the Vault data into another
cluster. I.e. when bootstrapping a Managed K8S cluster, you'll have to
hook it up to the correct vault instance right from the beginning (or
start over with fresh roots).