etcd-backup: don't insist on presence of section
etcd-backups are an optional feature so we should not force anyone to
update their config.toml after this MR has been merged. By commenting
out the `enabled` field in the `[.etcd-backup]` section,
update_inventory does not consider the variable as mandatory which is
fine because we can simply create a default in the role itself.