vault: add support for intermediate certificates in chain
To allow for intermediate certificates in the chain used for all the
Kubernetes-related certificates, we have to ensure that all certificates
include the full chain and that the trusted CA file is in fact the root.

Unfortunately, the latter requires messing with strings in ugly ways,
which is why I wanted to document this explicitly.