etcd-backup: Unpin chart version
We won't be able to follow the very short release cycles of the yaook
helm repository. As of 3 weeks after the previously pinned version, it
already wasn't available to me anymore. Helm charts do not have a
mechanism like "latest" tags or such. Therefore, this commit unpins the
etcd-backup chart version such that always the latest version will be
used.