Configurable Terraform root module path
The environment variable `TERRAFORM_MODULE_PATH` can be used to override
the path to the Terraform root module[^1]. This enables support for
custom Terraform-based environments that are not based on OpenStack by
providing custom resource definition files. If the environment variable
is unset or null, the default path `../terraform` is used.

[^1]: https://www.terraform.io/language/modules#the-root-module

Signed-off-by: Maximilian Nitsch <maximilian.nitsch@d3tn.com>