rook: Fix v1.2 and v1.3 CRD manifests
During the upgrade to Kubernetes v1.22 [1] the API versions have been
adjusted. This breaks the common (CRD) manifests with the following
error message:

> CustomResourceDefinition cephclusters.ceph.rook.io: resource definition
> validation error at spec: Additional properties are not allowed
> ''additionalPrinterColumns'', ''validation'', ''version'' were unexpected)'

Rook supports Kubernetes down to v1.6, but changing the manifests that
way seems to break the support. I did a diff on the example manifests of
the releases [2][3][4] and based on that I've decided to reset the
apiVersions. That fixed the problem.

[1] a2ef4349166f82ef270ad67e3e10583bf3cdaddf
[2] https://github.com/rook/rook/releases/tag/v1.2.3
[3] https://github.com/rook/rook/releases/tag/v1.3.11
[4] https://github.com/rook/rook/releases/tag/v1.4.9