vault: revamp support for importing existing clusters
Vault 1.11 now requires that the CA certificates have the crlSign bit
set; otherwise, it'll refuse to import them. In addition, with the new
multi-issuer setup, it is required that the cluster is *not*
pre-registered with Vault. Otherwise, the new keypair will only be
installed as a new issuer, not replacing the default issuer.