terraform: Fix migrate count for each - shellcheck issues
Enables the optional check: https://www.shellcheck.net/wiki/SC2312

To make shellcheck happy, the "lib.sh" has been sourced in the
"migrate-count-for-each.sh" script in [1]. However, that broke the
script because the path to "lib.sh" was faulty.

"migrate-count-for-each.sh" contains function definitions only itself.
It is sourced by the "apply-terraform.sh" script which as well sources the
"lib.sh" script, so all variables from it are available after invocation
and there is no need to source "lib.sh" in "migrate-count-for-each.sh",
from a functional point of view. It even breaks the functionality of
"lib.sh", because the variables will be faulty if lib.sh is sourced in
the migration script, as the "migrate-count-for-each.sh" script is in a
subdirectory of the other (usual) action scripts which will lead to
faulty paths due to [2]. This is the first case where a helper script
(located in actions/helper) uses variables from lib.sh and therefore it
is the first time that this falls on our feet.

I was not able to make shellcheck happy by
using directives [3] (i.e. by setting "source" or "source-path"), so
this commit disables the check SC2154 [4] as the functions do use
variables defined in "lib.sh" but we do not want to source it here.

[1] 05eac1f2f1e72f4810143b8b5090b6016c1eb50d
[2] https://gitlab.com/yaook/k8s/-/blob/e0fed55c607bd3b7573cdfedb8a9544c66fc4a89/actions/lib.sh#L2
[2] https://www.shellcheck.net/wiki/Directive
[3] https://www.shellcheck.net/wiki/SC2154