is-k8s-installed: Do not check for CNI config
As explained in #487, this check overshoots the mark a little. This role
is called "is-k8s-installed" and not "is-k8s-installed-and-set-up" :)

Rationale from the linked issue:
If stage 3 aborts after joining the node but before provisioning the CNI
configuration, is-k8s-installed will detect the node as k8s_not_installed.
That then causes a re-run of stage 3 to fail because a join is attempted once more.

These tests have been introduced in !526. Prior to that one, the test
were insufficient. However, now they are a bit over the top due to
the problem specified in #487.

Closes: #487