ci: k8s upgrade path: Do not install ingress for v1.19
The ingress chart requires a Kubernetes version >=1.20.0-0 and therefore
attempting to install it on a v1.19 cluster fails.

As we do not want to bounce back to a nginx-ingress chart version
supporting v1.19 (because k8s v1.19 is EOL for some time now...), this
commit disables ingress for the v1.19 cluster setup. It will be installed
after the cluster has been updated to v1.20 though.