ci: Remove timeouts
As there are no retries anymore, I do not see the point in forcing a
stage to fail after a certain amount of time. Also it is hard to set the
time limit here. Just one example: maybe the underlying ceph is slow
and the tests would succeed, but not within 30m. Also, who keeps an eye
on these timeout values with further tests and an increasing feature set?

One can cancel a job thanks to !470 if he foresees it to fail and wants
to free the resources to trigger another one.