CI: k8s-upgrades: Use helper script to patch config
In !450 we're reintroducing a CI stage which automatically tests the
Kubernetes upgrade path. After each upgrade to a newer Kubernetes
version, one would have to adjust the "k8s_version" in the config.toml.
We've previously were circumventing that by overwriting this variable
via the AFLAGS [1] mechanism before starting the next update
(see gitlab-ci.yaml [2]).

For Kubernetes version prior to v1.22, we do patch the CCM [3]. For
the following versions, we do not want to patch it. When using the
AFLAGS-mechanism in the Kubernetes CI upgrades, the following problem
would occur:
Ansible variables passed as extra vars (via "-e", so via our AFLAGS
mechanism), will ALWAYS take precedence [4]. As the Kubernetes version
is passed in that way in the CI, Ansible included the CCM patch tasks
despite the circumvention implemented in the upgrade playbook [5].
Also, (once again, because AFLAGS is used to set "k8s_version") the
versions mapped to K8s v1.21 will be rolled out which breaks the upgrade
to v1.22 [6] (!506).

This commit adds a simple python script which is used in the CI to patch
the existing config. That way, we do not have to worry about precedence
of variables in the CI passed via AFLAGS.

[1] https://yaook.gitlab.io/k8s/usage/environmental-variables.html#behavior-altering-variables
[2] https://gitlab.com/yaook/k8s/-/blob/51fdf196a94fb4d24f63041cb5a11c0bd0c65061/.gitlab-ci.yml#L210
[3] https://gitlab.com/yaook/k8s/-/blob/0e1113beaaf63a84dab0f150693892413f2fc79a/k8s-base/roles/connect-k8s-to-openstack/tasks/main.yaml#L5-8
[4] https://docs.ansible.com/ansible/latest/user_guide/playbooks_variables.html#understanding-variable-precedence
[5] https://gitlab.com/yaook/k8s/-/blob/0e1113beaaf63a84dab0f150693892413f2fc79a/k8s-base/k8s-upgrades/k8s_upgrade_to_1.22.yaml#L43
[6] fde5aad38841e4816ef80174457498d4a908cdb3