k8s-upgrades: Add intermediary cluster health verification
Please do also refer to the rationales stated in #491.

This commit adds a new role "cluster-health-verification". This role is
intended to be used for generic checks against a Kubernetes cluster
to ensure it's healthiness. In contrast to the smoke tests, these tasks
are not meant to be exhaustive.

That role is executed prior to the upgrade of EACH node.

As a first step, the role ensures that the rook-based Ceph cluster is
healthy. As mentioned in #491, if we do taint and upgrade the nodes to
fast, it may happen that the ceph cluster ends up in an unhealthy state.
I did also observe that when testing !450.

As the cluster health verification tasks of rook are on the KSL layer,
the upgrade.sh procedure needs to have access to the KSL roles and
the KSL variables have to be included where needed.

The role can be skipped by passing "-s" as flag to the upgrade.sh script.

Closes: #491