monitoring_v2: Allow label selectors for Prometheus
By default, the LCM-managed Prometheus will pick up any ServiceMonitor.

This commit adds the option to define labels in the config.toml which
will then be used by Prometheus to select ServiceMonitors.
This allows to narrow down the scope of the LCM-managed Prometheus and
allows for example to deploy multiple Prometheus which then can handle
different ServiceMonitors.