update_inventory: Add option to keep certain keys un-flattened
This commit adjusts the inventory updater such that it is possible to
keep certain keys un-flattened. The keys are hardcoded and need to be
defined per top level section. This is useful to e.g. define a list of
dictionaries in a section. Previously, each key would have been
flattened, destroying the original structure (a list of dicts).

This is needed for !578.