monitoring: Allow persistence for Grafana
This commit allows to Grafana to use persistent volumes [1]. When
migrating from non-persistence to persistence, a new Grafana pod will be
spawned (obviously as it bounds a persistent volume) and all data will
be migrated.

Why no simple "*_enabled" switch? Because there may be no default
storage class set in the Kubernetes cluster. Actually, we do not set a
default storage class with our LCM so there is a good chance there is
none set.In the case that there is no default storage class set, but a
simple "*_enabled" switch is set to true, the task will fail as the user
must provide a valid storage class name if there is no default.

[1] https://github.com/grafana/helm-charts/blob/main/charts/grafana/README.md?plain=1#L98