vault: do not use a shared container name for the management vault
This allows for separate vaults per repository. We persist the container
name in the repository, in order to not start ever and ever more
containers. We cannot use the repository directory or path as a reliable
identifier (even for initializing the persisted container name), because
repositories may swap locations in their lifetime, thus risking
conflict.

We now also run Vault as the current user. This means we have to give up
locking Vault contents in memory (because the process inside docker does
not have the privileges to enable the IPC_LOCK capability on the Vault
binary before startup), but that is OK as this is only supposed to be
used for development setups.