rook: Cluster Health verification, reduce waiting time
During rook upgrades and during Kubernetes upgrades (after !574),
we're checking the rook ceph cluster health by, among other things,
checking that all Pods in the rook-ceph namespace are up and running.

However, during upgrades Pods get restarted. If there is a Pod currently
in "Terminating" when this task is triggered, the task will run into
a timeout (after the default 120s) as that Pod will not meet the waiting
condition. This commit reduces the timeout to 10s. As the cluster health
verifications are run in a loop, that reduces unnecessary waiting time
and the next check is executed earlier.