vault: fix use of approle credentials
As it turns out, ANSIBLE_HASHI_VAULT_* does not work for the auth
method, auth mount point, role id and secret id fields. Because of that
we have to pass the things explicitly on all invocations and cannot use
an `environment` block for this.

To mitigate that, we introduced the most commonly used things into
defaults in `vault-config`.