vault: improve import flow
When I was about to fully document the two possibilities to import an
existing cluster into vault, I realized that the words "remove the
existing issuers" are much more easily typed than explained.

To fix that, the import.sh script was changed to support this mode
explicitly, avoiding the ambiguity and trickery otherwise necessary.