local-storage-controller: fix storage class map for custom sc names
The default storage class name for the local-storage-controller (a.k.a.
static local storage) is `local-storage`. The ability to customize this
existed since forever (since the introduction of
local-storage-controller in 1a018ba), but it apparently never did work
correctly.

Being able to do this correctly is now urgent since the introduction of
the dynamic local storage in d054443, as both default to the name
`local-storage` for their storageclass.

To make this possible, we need to tell the local storage controller
about its storage class name correctly.