local-storage-controller: make actually upgradable
Previously, changing the version would be impossible since that'd
attempt to change the DaemonSet selector, which is immutable.

As a fix, we freeze the /version field in the pod labels and introduce
an /actually-version field for informational purposes.

Also, the version wasn't even *used* in the main container image...
that's now fixed, too.