tf: skip count->foreach migration if no state exists
`terraform_migrate_foreach` assumes the existence of a terraform
state file (what else should it work with otherwise?). This commit
adds a guard to check if that's the case before invoking the
function.

Note: we make a strict assumption on the path of the statefile here. I'm
not aware, though, of any *relevant* cluster repo where this assumption
does not hold.