SSH: Re-add host key checking
This commit adds SSH host key checking. We're aware that TOFU is not
ideal. This is just an intermediate step, SSH CA handling will follow.

The known hosts file is located in the inventory at
"intventory/.etc/ssh_known_hosts". However, until now we did not make
use of it and disabled host key checking, which is a security flaw.

The action scripts and the Ansbile configuration have been adjusted to
use host key verification (which is the proposed default anyway).

The already present role "ssh-known-hosts" has been adjusted to usefully
manage the known hosts file and to fail if a host key changed.

We do have to allow host key changes directly after the devsec hardening
as that role tinkers with the host keys. In especially, it currently
replaces the default 3072-bits RSA key-pair with a 4096-bits big one.
However, to avoid race condition on the known hosts file, we're doing
that in serial to be sure the outcoming known hosts file is correct and
consistent.

The "ssh-known-hosts" role is now a dependency of the "detect-user" role
because we want to make sure that the former is already executed before
the later.

Force allowing host key changes can be done by passing:
-e allow_host_key_changes=true

Closes: #494