destroy.sh: Fix Wireguard interface tear down
If one destroys its cluster and the wireguard interface is not up, the
destroy.sh script fails and pending tasks are not executed.

This commit adds a condition which checks if the wireguard interface is
up before attempting to tear it down. While one can check a wireguard
interface via "wg show INTERFACE", we do not want to use that for this
condition as that requires unnecessary (at least for our intention here)
privileges.