Create upgrade routine for chart kube-prom-stack chart
We cannot upgrade chart versions directly because we need to update CRDs
as well [1].

This commit creates an upgrade routine which conditionally applies
version specific CRDs along with upgrading the charts.
A jump between major version is possible, therefore this
commit iteratively applied the CRDs of the jumped version as well.
This commit takes care of upgrade between version 33.x to version 39.x

An upgrade requires removing old prometheus admission controller of
previous version, Ref: https://github.com/helm/charts/issues/19928
Therefore, this commit also takes care of this

This commit doesn't use the Ansible k8s module to apply the CRDs because
of an issue regarding the 'altermanagerconfigs' CRD.
This yaml contains an enum `- =` which is not processed by pyyaml and it
throws an error [2]. Therefore the from_yaml filter fails and we cannot
read the yaml file. There is also a pending pull request regarding the
update of this CRD [3].
Once this request is approved, This commit can be updated to use the
Ansible k8s module.

There is currently no equivalent of `helm show chart <chart name>` in
Ansible helm module. Therefore instead of using helm module we use
command module to find the appverison of a helm chart.

[1] https://github.com/prometheus-community/helm-charts/blob/main/charts/kube-prometheus-stack/README.md#upgrading-chart
[2] https://github.com/yaml/pyyaml/issues/89
[3] https://github.com/prometheus-operator/prometheus-operator/pull/4897