smoke-tests: Unify namespace
Instead of creating each test in it's own namespace, the namespace is
now standardized to be yaook-k8s-tests-2cdd393b-2f8a-4ffa-8bd8-4e636e5154ec

The reason for that is that using only one namespace makes it easier to
allow(-list) all smoke tests if there e.g. is a policy engine blocking
the creation of arbitrary namespaces.