smoke-tests: Use unprivileged containers
The smoke tests do not necessarily need privileged containers.
The privileged mode disables most security mechanisms and most not be
allowed. This reduces the risk that an adversary is able to escape out
of the container and getting access to the host Kubernetes node.

For the (loadbalanced-)service checks we now have to use an nginx image
which runs nginx as non-root, unprivileged user (related: [1]).

[1] https://gitlab.com/yaook/images/k8s-nginx/-/merge_requests/3/pipelines