etcd-backup: Enforce specification of CA cert bundle
If no CA cert bundle is specified, verification will be turned off [1][2].
This commit adjusts the etcd-backup-configuration template such that it
is clear that certs must be provided. Everything else would be a security
flaw. It also adjusts the etcd-backup role such that the specified certs
are passed through to the etcd-backup helm chart.

[1] https://gitlab.com/yaook/images/backup-shifter/-/blob/eed26006b4d5d69ec34af9061de5c72c1a983045/shifters/s3_upload#L29-30
[2] https://gitlab.com/yaook/operator/-/blob/ccabb7edfe25368548b5e54402729cfcee339202/yaook/helm_builder/Charts/etcd-backup/templates/configmap.yaml