CI: k8s-upgrades: Remove kube-router jobs
As we use a pretty old kube-router version and want to drop support for
it in general rather soon, this commit drops the jobs which test the
Kubernetes upgrade path with a kube-router Cluster. I doubt that the
version we're deploying is compatible with recent Kubernetes versions
anyway.

Prior to upgrade Kubernetes, you should exchange kube-router with calico.