Fix for_each migration broken on undeclared variable
When undeclared variables exist in [terraform] section the
'terraform console' command raises a warning and writes it to the
stdout and not to the stderr. This is a problem for a migration script,
because it uses stdout of that terraform command. For now, warnings
cannot be silenced in terraform so this commit just adds 'tail -1' to
get actual output on the last line of 'terraform console' stdout.

Fixes yaook/k8s#501

Signed-off-by: Roman Hros <roman.hros@dnation.cloud>