Tools: patch-config: Allow subtraction
This commit extends the patch-config script so that it is now possible
to subtract the specified JSON from the config. Previously, it was only
possible to merge and impossible to delete entries. However, deleting
entries may be handy when defining variables only for a certain period
of time after which the default shall be used again.

The script is moved to the "tools/" directory. Although it is currently
used only there, it is not limited to it.