lbaas: Ensure that secret is base64 encoded
I frequently forget to set a shared secret in the ch-k8s-lbaas section.
Since the variable has a default, the rollout will not fail but lbaas
itself cannot run because it expects the secret to be a base64-encoded
string. This commit adds a check and fails if the condition is not met.

Note: Other options are thinkable, like letting update_inventory roll
the dice or using vault once it's been introduced. We can implement them
later but I'd like to see this particular issue resolved now.