tf: Make monitoring_use_thanos optional
This commit adds a default (False) for the usage of thanos which is
declared as an optional variable in the config template. As a matter of
fact it wasn't optional before because update_inventory wrote `null`
into terraform's config.tfvars.json (instead of a boolean) which lead to
an invalid expression.