monitoring: Relabel and test blackbox endpoints
After upgrading to helm chart version v7.0.0, we can relabel metrics
and endpoints of the blackbox exporter. This is in turn allows us to
check in the smoke tests to check if the blackbox exporter is working
for each target. Beforehand, we simply deployed the blackbox exporter
but didn't actually check if it is running and working.

Closes: #492