monitoring: Hardcode blackbox release name
This variable is not reflected in the config.toml anyway. Hardcoding
it makes our lives easier, as the "job" label of a target depends on the
release name. Hardcoding it easens variable handling for us in e.g. the
blackbox smoke tests. If one really needs to make this configurable,
we can do that in a follow up MR in the future.