Improve POSIX compliance of remote scripts
POSIX does not guarantee /bin/bash to exist. It mandates the presence of
`env`, however, which can be used to find bash.

This commit changes the shebang of most remote scripts to use
/usr/bin/env.
Exceptions are /bin/sh which is POSIX compliant in itself and cases in
which the interpreter needs to be called with additional parameters, eg.
'/usr/sbin/nft -f', which is not possible using `env`.