kubeadm: Explicitly set cgroupdriver to systemd
This commit adjusts the kubeadm cluster initialization configuration
file such that systemd is explicitly set as cgroup driver.

systemd is the default cgroup driver since Kubernetes v1.22 [1].
The previous default was "cgroupfs". We never did explicitly specify
the cgroup driver for kubelet, but only in the docker configuration [3].
kubelet seems to have some auto detection mechanism basing on the docker
configuration.

As we now also want to support containerd as CRI, that mechanism seems
not to catch. The consequence is, that kubeadm can't initialize a new
cluster with Kubernetes versions below v1.22 (because the wrong default
is used in that versions). Therefore, this commit explicitly sets
systemd as cgroup driver.

[1] https://kubernetes.io/docs/tasks/administer-cluster/kubeadm/configure-cgroup-driver/#configuring-the-kubelet-cgroup-driver
--cgroup-driver string     Default: cgroupfs
[2] https://kubernetes.io/docs/reference/command-line-tools-reference/kubelet/
[3] https://gitlab.com/yaook/k8s/-/blob/9f7264803f9f2b9ffe08228323f5e40f05dea26d/k8s-base/roles/docker/templates/daemon.json.j2#L17