gitlab-ci: Retry non-manual jobs two times
It may happen that a job fails due to hickups in the underlying
infrastructure or network problems. That is annoying because the whole
pipeline then has to be re-triggered. Our CI pipeline is very time and
resource consuming, so we want to avoid re-running it completely for
such failures. However, we also do not want to manually force merge in
such cases, as that may introduce some regressions.

Therefore, this commit lets the non-manually triggered jobs retry two
times. This does affect merge pipelines, scheduled pipelines as well as
pushes on the main branch(es).