CI: Add docker k8s upgrade job, rename jobs
As there are productive clusters which use docker and also GPU clusters
must use docker, we do need to test the Kubernetes upgrade path with docker
in the CI. This will add another job to the CI matrix (and quite a time-
consuming one). However, as we now retry non-manual jobs, there is at
least an increased chance that MR trains do not have to be re-triggered
manually so often anymore.

This commit does also adjust the naming scheme, because it really
confused me. Every job uses rook, so no need to add it to the name.
Every job now uses calico as CNI, so also no need to include that.
However, the interesting part is which CRI is used.

Closes: #511