config: Purge legacy converter
The legacy converter for the configuration was a helper script which was
needed to update the configuration files of clusters prior to !288.
In !288 the inventory handling has been refactored.
The script has been introduced in 382d6d75b64ddbd03b48f3fdde95d9e27f9ff1cd
and is not needed anymore.

Closes: #420