rook: Add support for rook v1.7.11 w/ Ceph v16.2.6
Enhancements of rook are described in [1].

The upgrade routine implementation is kind of straight-forward and how
to implement a new rook version is described in [2]. As there are no
significant changes, no problems occured when adjusting the templates
from [3].

The ceph images have been moved to quay.io, therefore we need to patch
the CephCluster object depending on the version to be deployed.

Although support has been added for Ceph-cluster helm charts, we're not
going to use them, yet.

[1] https://blog.rook.io/rook-v1-7-storage-enhancements-6ae647aa5d97
[2] https://gitlab.com/yaook/k8s/-/blob/02312b1ce3d4b98d149ec92d3d1ea6d3204431c5/docs/src/managed-services/rook/upgrades.md#addingimplementing-support-for-a-new-rookceph-release-to-managed-k8s
[3] https://github.com/rook/rook/releases/tag/v1.7.11