CI: Do not retry manual jobs
The retry value is inherited from the template (of the auto-triggered
jobs). We have to explicitly set it to zero for manual jobs, because we
do not want that manual jobs get retried (as we see it as one shot check
if everything is alright before doing an in-depth review).