Add option: gitlab-managed terraform http backend
If enabled, terraform state is stored on Gitlab,
and the local copy of it is updated.

Backend of Terraform on Gitlab gets benefits,
such as shared storage, locking, versioning,
and encryption of state file.

Required env variables:
- gitlab_backend
- gitlab_base_url
- gitlab_project_id
- gitlab_state_name
- TF_HTTP_USERNAME (username on Gitlab)
- TF_HTTP_PASSWORD (token for Gitlab)

When switching from local to http backup, the
local backup file `terraform.tfstate.backup` could
be out-to-date until the switching from http backup
to local backup.