bbe: Add a tcp prober
This commit adds a simple tcp prober to the configuration of the
blackbox exporter. It can be used in situations where you don't care
about particular HTTP status codes and just want to see if a (TCP)
service is reachable at all.

Note: If we need more modules than either we can extend this list or let
the LCM define a preconfigured secret from which the bbe can fetch its
config (so its config is under direct control of the end user and
changes don't have to be merged into the LCM).