thanos: Test endpoints and support common labels
The LCM allows you to specify a set of common labels which is added to
all monitoring resources and which also applies to the
ServiceMonitor-selector of Prometheus. The latter means that if
`common_labels` is not empty, then Prometheus will only scrape
ServiceMonitors (or "targets") if they have these labels set.

This commit sets these labels correctly for all thanos components.
Sidenote: we're still using the thanos jsonnet setup because there's no
official helm chart and we didn't find the time yet to evaluate the
community ones.

Furthermore the test suite now actively probes that Prometheus scrapes
all thanos metrics endpoints (i.e., the metrics of thanos itself).