Ensure OS credentials are sourced in lbaas, CCM
As the comment in the config.template.toml states, terrible things will
happen if certain tasks are executed inside a cluster that runs on top
of OpenStack. This commit adds `check-openstack-credentials` as a
gatekeeper role which, when it passes, gives us sufficient confidence
that the required credentials are sourced. It guards two of the
three dangerous spots which requer sourced credentials, namely
ch-k8s-lbaas and connect-k8s-to-openstack. The third, thanos, is dealt
with in the next commit.

A note on the implementation: One could have checked for the existence
of certain variables but please mind that we support both application
credentials and user/password authentication. I deemed it easier (and
more sustainable because it even validates the credentials themselves)
to throw them against the keystone service and see whether any token
pops out. That check can be skipped, of course.