k8s-upgrades: Add calico update to upgrade procedure
Although we allow users to pin the calico version, under normal
circumstances a user will use the calico version defined in the
"k8s-config" role. The calico version is mapped to the Kubernetes
version. Here is the problem which occurs with this:

During a Kubernetes upgrade to a new Kubernetes version, we do not
update components which versions are mapped to the configured Kubernetes
version. Despite that, we're checking i.e. if the configured calico
version matches the Kubernetes version in stage3. This leads to failure
on a consecutive stage3 runs after an Kubernetes upgrade (if no
disruptions are allowed). That behavior is confusing for normal users.

To solve this issue, this commit adds the calico update to the
Kubernetes upgrade procedure.