k8s-upgrades: Add CCM/Cinder updates to upgrade procedure
The versions of the services and modules which connect the Kubernetes
cluster to the underlying OpenStack (if we're running on OpenStack)
are mapped to the configured Kubernetes version. Therefore, we need to
update these components in the upgrade procedure. Otherwise they would
be updated on a consecutive stage3 run, which may lead du unwanted
effects and is probably unexpected by unaware users.

For k8s versions < v1.22 we haven't touched these component versions...
For k8s version = v1.22 we update these components prior to the actual
update, because that's a hard requirement (see eb3a2bd268a2ef7fde363028c96b0f07348d0585)
For versions >= v1.23 we want to update these components after the
Kubernetes update itself.

---
PS: The main reason because the CCM update is not included prior to
v1.22 is that in that mapped versions we patch the kube controller
manager [0][1] with this [2] script, but the path to the script differs
within kubernetes upgrade plays (because these playbooks are one level
below the stage3 playbook) and therefore the "connect-k8s-to-openstack"
role fails for k8s versions < v1.22 as the script can't be found.

[0] https://gitlab.com/yaook/k8s/-/blob/598d6cacde8d3cdcd41274dfc37f397c47692d61/k8s-base/roles/connect-k8s-to-openstack/tasks/main.yaml#L3-8
[1] https://gitlab.com/yaook/k8s/-/blob/598d6cacde8d3cdcd41274dfc37f397c47692d61/k8s-base/roles/connect-k8s-to-openstack/tasks/patch_controller_manager.yaml#L2
[2] https://gitlab.com/yaook/k8s/-/blob/598d6cacde8d3cdcd41274dfc37f397c47692d61/tools/patch_config.py